Split out record and replay tests from gcsblob_test
